requirements.txt
botcity-framework-web>=0.9.0,<1.0
botcity-maestro-sdk>=0.3.3,<1.0
botcity-utils>=0.1.2
botcity-excel-plugin>=0.1.7

# Uncomment to use botcity-crawler-plugin
# botcity-crawler-plugin

# Uncomment to use Pandas
# pandas
